requirements.txt
Flask>=2.0.0
Flask-Limiter>=2.0.0
Flast-Talisman>=0.8.0
google-cloud-storage>=1.43.0
google-cloud-secret-manager>=1.1.0
